# Speed up parsing units with scale factors