--- a/docs/changes/units/16813.perf.rst
+++ b/docs/changes/units/16813.perf.rst
@@ -1,0 +1,1 @@
+Parsing units with scale factors is now up to 50% faster.