### Description

Setup:
```
ipython -i -c 'from astropy import units as u'
```
On current `main`:
```
Python 3.10.12 (main, Jul 29 2024, 16:56:48) [GCC 11.4.0]
Type 'copyright', 'credits' or 'license' for more information
IPython 8.26.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: %timeit u.Unit("1000m", format="cds")
31.9 μs ± 1.29 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [2]: %timeit u.Unit("10**3 m", format="fits")
38 μs ± 92.4 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [3]: %timeit u.Unit("10**3 m", format="generic")
40.1 μs ± 1.94 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [4]: %timeit u.Unit("10**3 m", format="ogip")
38.4 μs ± 212 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [5]: %timeit u.Unit("10**3 m", format="vounit")
38.2 μs ± 150 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

With the patch here:
```
In [1]: %timeit u.Unit("1000m", format="cds")
16.6 μs ± 80.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [2]: %timeit u.Unit("10**3 m", format="fits")
26.6 μs ± 1.18 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [3]: %timeit u.Unit("10**3 m", format="generic")
24.1 μs ± 110 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [4]: %timeit u.Unit("10**3 m", format="ogip")
24.6 μs ± 395 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [5]: %timeit u.Unit("10**3 m", format="vounit")
25.9 μs ± 63.2 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

- [x] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.